requirements.txt
absl-py==0.9.0
astor==0.8.1
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cycler==0.10.0
gast==0.2.2
google-api-core==1.14.3
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-speech==1.3.1
google-cloud-texttospeech==0.5.0
google-pasta==0.2.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
h5py==2.10.0
idna==2.8
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.1.2
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.1
protobuf==3.11.1
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyAudio==0.2.11
pycairo==1.19.1
pycparser==2.19
pydub==0.23.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
six==1.13.0
sklearn==0.0
sounddevice==0.3.14
SQLAlchemy==1.3.13
tensorboard==2.1.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==1.0.1
wrapt==1.12.1
